+++ src/tests/EntityGraphQL.Tests/QueryTests/ArgumentTests.cs
@@ -57,7 +57,7 @@ namespace EntityGraphQL.Tests
                 }"
             };
             var result = schema.ExecuteRequest(gql, new TestDataContext(), null, null);
-            Assert.Equal("Field error: user - 'user' missing required argument 'id'", result.Errors[0].Message);
+            Assert.Equal("Field 'user' missing required argument 'id'", result.Errors[0].Message);
         }
 
         [Fact]
@@ -73,7 +73,7 @@ namespace EntityGraphQL.Tests
                     }"
             };
             var result = schema.ExecuteRequest(gql, new TestDataContext(), null, null);
-            Assert.Equal("Field error: user - 'user' missing required argument 'id'", result.Errors[0].Message);
+            Assert.Equal("Field 'user' missing required argument 'id'", result.Errors[0].Message);
         }
 
         [Fact]