+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Addition changes
 - Performance of compiling and building expressions has been improved - it is about 2 times faster. Note this is just building the expressions, not executing them which triggers your services/EF/etc
 - Reduction in memory allocation when compiling queries by around 30%
 - You can no add/define mutation methods as delegates/anonymous functions etc.
+- You can now use `[Range]` & `[StringLength]` attributes on your arguments for more validation options
 
   Fixes